Better weighted selection test.
