+++ cmd/block-generator/generator/generate.go
@@ -183,26 +183,6 @@ func (g *generator) WriteGenesis(output io.Writer) {
 	output.Write(json.Encode(gen))
 }
 
-func weightedSelection(weights []float32, options []interface{}) (selection interface{}, err error) {
-	if len(weights) != len(options) {
-		err = fmt.Errorf("number of weights must equal number of options: %d != %d", len(weights), len(options))
-		return
-	}
-
-	selectionNumber := rand.Float32()
-	total := float32(0)
-	for i := 0; i < len(weights); i++ {
-		if selectionNumber - total < weights[i] {
-			selection = options[i]
-			return
-		}
-		total += weights[i]
-	}
-
-	err = outOfRangeError
-	return
-}
-
 func getTransactionOptions() []interface{} {
 	return []interface{}{paymentTx, assetTx}
 }
